pyproject.toml
[project]
name = "events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "kafka-python>=2.3.1",
    "locust>=2.43.4",
]
